Remove support for Java 11 (#2308)
1. Why:
Jetty 12 requires a minimum of Java 17, so Java 11 must be dropped to facilitate this upgrade.
The current version of Jetty in cruise-control is 9.4.56.v20240826. This version contains two CVEs
CVE-2024-13009 - patched in 9.4.57.v20241219
CVE-2024-6763 - patched in 12.0.12
So, Jetty 12 is need to patch CVE-2024-6763.

2. What:
Remove support support for Java 11
Raise minimum JDK to 17 in builds, docs, and workflows

- [x] refactor

Signed-off-by: JvD_Ericsson <jeff.van.dam@est.tech>